web/requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.3
argparse==1.2.1
itsdangerous==0.24
wsgiref==0.1.2
Flask-SQLAlchemy==2.1
SQLAlchemy-migrate>=0.10.0
isc_dhcp_leases==0.5.3
flask-login>=0.3.2
astral>=0.9
